if else bind (sorta)

+++ src/backend/cpp/runtime/cppruntime.hpp
@@ -248,7 +248,7 @@ public:
     }
 };
 
-typedef uint64_t None;
+typedef uintptr_t None;
 
 #define MAX_BSQ_INT ((int64_t(1) << 62) - 1)
 #define MIN_BSQ_INT (-(int64_t(1) << 62) + 1)
